requirements.txt
absl-py>=2.1.0
altair>=5.0.0
astor>=0.8.1
attrs>=23.1.0
backcall>=0.2.0
base58>=2.1.1
bleach>=6.0.0
blinker>=1.6.2
boto3>=1.26.0
botocore>=1.29.0
cachetools>=5.3.0
certifi>=2023.7.22
chardet>=5.1.0
click>=8.1.3
colorama>=0.4.6
decorator>=5.1.1
defusedxml>=0.7.1
docutils>=0.19
entrypoints>=0.4
enum-compat>=0.0.3
future>=0.18.3
gast>=0.5.3
google-auth>=2.20.0
google-auth-oauthlib>=1.0.0
google-pasta>=0.2.0
grpcio>=1.60.0
gunicorn>=20.1.0
h5py>=3.9.0
idna>=3.4
importlib-metadata>=6.0.0
ipykernel>=6.29.0
ipython>=8.17.2
ipython-genutils>=0.2.0
ipywidgets>=8.0.4
jedi>=0.19.1
Jinja2>=3.1.2
jmespath>=1.0.1
jsonschema>=4.19.0
jupyter-client>=8.4.0
jupyter-core>=5.5.0
Keras-Applications>=1.0.8
Keras-Preprocessing>=1.1.2
Markdown>=3.4.4
MarkupSafe>=2.1.3
mistune>=3.0.1
nbconvert>=7.11.0
nbformat>=5.9.2
notebook>=7.0.6
numpy>=1.23.0
oauthlib>=3.2.2
opt-einsum>=3.3.0
packaging>=23.2
pandas>=2.0.3
pandocfilters>=1.5.1
parso>=0.8.3
pathtools>=0.1.2
pickleshare>=0.7.5
Pillow>=10.0.0
prometheus-client>=0.18.0
prompt-toolkit>=3.0.41
protobuf>=4.23.4
pyasn1>=0.5.1
pyasn1-modules>=0.3.0
pydeck>=0.8.0
Pygments>=2.16.1
pyparsing>=3.1.1
pyrsistent>=0.20.0
python-dateutil>=2.8.2
pytz>=2023.3
pywinpty>=2.0.11
pyzmq>=25.1.1
requests>=2.31.0
requests-oauthlib>=1.4.0
rsa>=4.9
s3transfer>=0.6.2
scipy>=1.11.3
Send2Trash>=1.8.2
six>=1.16.0
streamlit>=1.30.0
tensorboard>=2.15.0
tensorflow>=2.13.0
tensorflow-estimator>=2.13.0
termcolor>=2.4.0
terminado>=0.18.0
testpath>=0.6.0
toml>=0.10.2
toolz>=0.12.0
tornado>=6.3.3
traitlets>=5.11.2
tzlocal>=5.2
urllib3>=2.0.7
validators>=0.22.0
watchdog>=3.0.0
wcwidth>=0.2.9
webencodings>=0.5.1
Werkzeug>=3.0.1
widgetsnbextension>=4.0.9
wrapt>=1.16.0
zipp>=3.17.0
